chore: use swapper selectors (#4078)
