+++ src/components/Trade/TradeInput.tsx
@@ -39,7 +39,12 @@ import {
   selectPortfolioCryptoHumanBalanceByFilter,
 } from 'state/slices/selectors'
 import { useAppSelector } from 'state/store'
-import { selectSlippage } from 'state/zustand/swapperStore/selectors'
+import {
+  selectCheckApprovalNeededForWallet,
+  selectQuote,
+  selectSlippage,
+  selectSwapperSupportsCrossAccountTrade,
+} from 'state/zustand/swapperStore/selectors'
 import { useSwapperStore } from 'state/zustand/swapperStore/useSwapperStore'
 import { breakpoints } from 'theme/theme'
 
@@ -72,7 +77,7 @@ export const TradeInput = () => {
   const updateSelectedBuyAssetAccountId = useSwapperStore(
     state => state.updateSelectedBuyAssetAccountId,
   )
-  const activeQuote = useSwapperStore(state => state.activeSwapperWithMetadata?.quote)
+  const activeQuote = useSwapperStore(selectQuote)
   const fees = useSwapperStore(state => state.fees)
   const slippage = useSwapperStore(selectSlippage)
   const updateFees = useSwapperStore(state => state.updateFees)
@@ -105,22 +110,16 @@ export const TradeInput = () => {
   const updateSellAmountCryptoPrecision = useSwapperStore(
     state => state.updateSellAmountCryptoPrecision,
   )
+  const swapperSupportsCrossAccountTrade = useSwapperStore(selectSwapperSupportsCrossAccountTrade)
+  const checkApprovalNeeded = useSwapperStore(selectCheckApprovalNeededForWallet)
 
-  const {
-    checkApprovalNeeded,
-    getTrade,
-    getSupportedSellableAssets,
-    getSupportedBuyAssetsFromSellAsset,
-    swapperSupportsCrossAccountTrade,
-  } = useSwapper()
+  const { getTrade, getSupportedSellableAssets, getSupportedBuyAssetsFromSellAsset } = useSwapper()
   const translate = useTranslate()
   const history = useHistory()
   const mixpanel = getMixPanel()
   const borderColor = useColorModeValue('gray.100', 'gray.750')
   const { handleSubmit } = useFormContext()
-  const {
-    state: { wallet },
-  } = useWallet()
+  const wallet = useWallet().state.wallet
   const tradeAmountConstants = useGetTradeAmounts()
   const { assetSearch } = useModal()
   const { handleAssetClick } = useTradeRoutes()
@@ -307,7 +306,8 @@ export const TradeInput = () => {
           [compositeSellAsset]: sellAmountCryptoPrecision,
         })
       }
-      const isApprovalNeeded = await checkApprovalNeeded()
+      if (!wallet) throw new Error('No wallet available')
+      const isApprovalNeeded = await checkApprovalNeeded(wallet)
       if (isApprovalNeeded) {
         history.push({ pathname: TradeRoutePaths.Approval })
         return
@@ -333,6 +333,7 @@ export const TradeInput = () => {
     sellAsset,
     swapperName,
     updateTrade,
+    wallet,
   ])
 
   const onSellAssetInputChange: TradeAssetInputProps['onChange'] = useCallback(
