+++ src/components/Trade/hooks/useFeesService.tsx
@@ -3,6 +3,7 @@ import { useEffect } from 'react'
 import { getFormFees } from 'components/Trade/hooks/useSwapper/utils'
 import { selectFeeAssetById } from 'state/slices/assetsSlice/selectors'
 import { useAppSelector } from 'state/store'
+import { selectQuote } from 'state/zustand/swapperStore/selectors'
 import { useSwapperStore } from 'state/zustand/swapperStore/useSwapperStore'
 
 /*
@@ -12,7 +13,7 @@ The only mutation is on Swapper State's fees property.
 export const useFeesService = () => {
   // Selectors
   const activeTradeSwapper = useSwapperStore(state => state.activeSwapperWithMetadata?.swapper)
-  const activeQuote = useSwapperStore(state => state.activeSwapperWithMetadata?.quote)
+  const activeQuote = useSwapperStore(selectQuote)
   const sellAsset = useSwapperStore(state => state.sellAsset)
   const sellFeeAsset = useAppSelector(state =>
     selectFeeAssetById(state, sellAsset?.assetId ?? ethAssetId),
